build.sh changes
* add SM build script and separate Linaro from SM
* clean the build scripts and add git repos for toolchains
* update zip.sh to make it compatible
* don't call zip.sh unless instructed
* I will merge the scripts in the near future

+++ zip.sh
@@ -2,10 +2,17 @@
 zipname="big-bum"
 version=$(cat .version)
 cd zip/
-cp ../boot-L-v"$version".img boot.img
-zip -q -r -9 "$zipname"-L-v"$version".zip *
-mv *.zip ../
-rm *.img
-if [ -e ../"$zipname"-L-v"$version".zip ]; then
-	echo "zip made"
+if [ $1 == "Linaro" ]; then
+	cp ../boot-L-v"$version"-Linaro.img boot.img
+	zip -q -r -9 "$zipname"-L-v"$version"-Linaro.zip *
+	rm *.img
+	mv *.zip ../
+	echo "Linaro zip made"
+fi
+if [ $1 == "SM" ]; then
+	cp ../boot-L-v"$version"-SM.img boot.img
+	zip -q -r -9 "$zipname"-L-v"$version"-SM.zip *
+	rm *.img
+	mv *.zip ../
+	echo "SM zip made"
 fi
